+++ packages/ui-components/web/src/index.ts
@@ -26,5 +26,6 @@ export { Tab, Tabs } from './Components/Tabs';
 export { default as Loader } from './Components/Loader';
 export { default as MediaItemErrorMsg } from './Components/MediaItemErrorMsg';
 export { default as TextSearchInput } from './Components/TextSearchInput';
+export { default as Separator } from './Components/Separator';
 
 // export { getSelectionStyles } from './Utils/inlineStyleUtils';