+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 ## 8.44.2 (Jun 14, 2021)
 ### :house: Internal
 - `editor-common`
-  - [#2617](https://github.com/wix-incubator/rich-content/pull/2617) moved textSearchInput component into ui-components folder
+  - [#2617](https://github.com/wix-incubator/rich-content/pull/2617) moved textSearchInput/Separator components into ui-components folder
 - `plugin-commons`
   - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved SettingsMobileHeader/UrlInputModal components into ui-components folder
 - `tiptap-editor`